+++ src/common/constants.ts
@@ -1,14 +1,25 @@
 import type { BasisTheoryServicesBasePathMap } from '@/types/sdk';
 
 const API_KEY_HEADER = 'BT-API-KEY';
+
 const BT_TRACE_ID_HEADER = 'bt-trace-id';
+
 const BT_IDEMPOTENCY_KEY_HEADER = 'bt-idempotency-key';
+
+const BT_EXPOSE_PROXY_RESPONSE_HEADER = 'BT-EXPOSE-RAW-PROXY-RESPONSE';
+
+const BT_TRANSACTION_ID_HEADER = 'BT-TRANSACTION-ID';
+
 const CONTENT_TYPE_HEADER = 'Content-Type';
+
 const USER_AGENT_HEADER = 'User-Agent';
+
 const CLIENT_USER_AGENT_HEADER = 'BT-CLIENT-USER-AGENT';
+
 const USER_AGENT_CLIENT = 'BasisTheoryJS';
 
 const DEFAULT_BASE_URL = `https://${process.env.API_HOST}`;
+
 const DEFAULT_ELEMENTS_BASE_URL = `https://${process.env.ELEMENTS_HOST}`;
 
 const CLIENT_BASE_PATHS: BasisTheoryServicesBasePathMap = {
@@ -23,6 +34,7 @@ const CLIENT_BASE_PATHS: BasisTheoryServicesBasePathMap = {
   proxies: 'proxies',
   proxy: 'proxy',
   sessions: 'sessions',
+  transactions: 'transactions',
 };
 
 const BROWSER_LIST = [
@@ -72,8 +84,10 @@ export {
   USER_AGENT_HEADER,
   CLIENT_USER_AGENT_HEADER,
   USER_AGENT_CLIENT,
+  BT_EXPOSE_PROXY_RESPONSE_HEADER,
   DEFAULT_BASE_URL,
   DEFAULT_ELEMENTS_BASE_URL,
   CLIENT_BASE_PATHS,
   BROWSER_LIST,
+  BT_TRANSACTION_ID_HEADER,
 };