+++ src/common/index.ts
@@ -2,3 +2,4 @@ export * from './constants';
 export * from './utils';
 export { BasisTheoryApiError } from './BasisTheoryApiError';
 export { BasisTheoryValidationError } from './BasisTheoryValidationError';
+export { HttpClientError } from './HttpClientError';