+++ src/service/types.ts
@@ -1,5 +1,4 @@
 import type { AxiosRequestConfig } from 'axios';
-import https from 'https';
 import type { ApplicationInfo } from '@/types/sdk';
 
 interface BasisTheoryServiceOptions
@@ -7,7 +6,6 @@ interface BasisTheoryServiceOptions
   apiKey: string;
   baseURL: string;
   appInfo?: ApplicationInfo;
-  httpsAgent?: https.Agent;
 }
 
 type RequestTransformers = Pick<